add scoring method comparing a syscallgroup with average group
The scoring system compares if there are new syscall introduced,
new files that have been read or written to, and the number of new PIDS in difference from average.

A new testing file was created to quickly test it out with some data from buffer overflow attack, and just normal usage.